feat(release): refactoring
the goal is to organize source code in different groovy files.
