+++ src/main/groovy/updateParentVersion.groovy
@@ -1,3 +1,5 @@
+package main.groovy
+
 def projects = [
         "gravitee-common",
         "gravitee-definition",
